requirements.txt
ultralytics>=8.3.2
matplotlib>=3.7.0
opencv-python>=4.7.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.4.0
PyYAML>=6.0
tqdm>=4.64.0
scipy>=1.10.0
flask>=2.0.0
requests>=2.25.0
pycocotools>=2.0.7
psutil>=5.9.0
albumentations>=2.0.0
# Optional dependencies for advanced features
# onnx>=1.14.0
# onnxruntime>=1.15.0
# gradio>=4.0.0 